Steganography/requirements.txt
asgiref==3.8.1
certifi==2024.12.14
charset-normalizer==3.4.0
Django==5.1.4
djangorestframework==3.15.2
idna==3.10
numpy==2.2.0
pillow==11.0.0
requests==2.32.3
sqlparse==0.5.3
typing_extensions==4.12.2
urllib3==2.3.0
